requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.3.0 # pyup: != 11.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==25.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.9.0  # https://github.com/evansd/whitenoise
redis==6.2.0  # https://github.com/redis/redis-py
hiredis==3.2.1  # https://github.com/redis/hiredis-py
celery==5.5.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.8.1  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
uvicorn[standard]==0.35.0  # https://github.com/encode/uvicorn
uvicorn-worker==0.3.0  # https://github.com/Kludex/uvicorn-worker

# Django
# ------------------------------------------------------------------------------
django==5.1.11  # pyup: < 5.2 # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.9.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.4  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2025.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==6.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.16.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.7.0  # https://github.com/adamchainz/django-cors-headers
django-filter==24.3  # https://github.com/carltongibson/django-filter
# DRF-spectacular for api documentation
drf-spectacular==0.28.0  # https://github.com/tfranzel/drf-spectacular
django-webpack-loader==3.2.1  # https://github.com/django-webpack/django-webpack-loader
# DRF simple-jwt for stateless authentication
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
djoser==2.3.1

# Document parsing (CV extraction)
pdfminer.six==20231228  # robust PDF text extraction (no AI)
pypdf==5.0.1  # fallback PDF text extraction (replaces deprecated PyPDF2)
pdf2image==1.17.0  # rasterize PDF pages for OCR fallback
pytesseract==0.3.13  # Tesseract OCR bindings

# Database URL parser
dj-database-url==2.3.0  # https://github.com/jazzband/dj-database-url

# Real-time Notifications
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.0.0
